test: ensure topology functionality
Verifies the topology functionality still works as expected.
The REST controller tests adjust to mocking the service layer instead of implementation details.
The former tests to ensure broker client and topology manager move to a new service layer test for topology services.
