+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/configuration/RestApiBrokerGatewayDisabledTest.java
@@ -44,6 +44,6 @@ public class RestApiBrokerGatewayDisabledTest extends RestApiConfigurationTest {
         .expectStatus()
         .isNotFound();
 
-    verifyNoInteractions(topologyManager);
+    verifyNoInteractions(topologyServices);
   }
 }
