+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/configuration/RestApiDefaultConfigurationTest.java
@@ -56,6 +56,6 @@ public class RestApiDefaultConfigurationTest extends RestApiConfigurationTest {
         .expectStatus()
         .isOk();
 
-    verify(topologyManager).getTopology();
+    verify(topologyServices).getTopology();
   }
 }
