+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/configuration/RestApiDisabledTest.java
@@ -44,6 +44,6 @@ public class RestApiDisabledTest extends RestApiConfigurationTest {
         .expectStatus()
         .isNotFound();
 
-    verifyNoInteractions(topologyManager);
+    verifyNoInteractions(topologyServices);
   }
 }
