+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/StatusControllerTest.java
@@ -7,21 +7,18 @@
  */
 package io.camunda.zeebe.gateway.rest.controller;
 
-import static org.mockito.Mockito.verify;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.SERVICE_UNAVAILABLE;
 
-import io.camunda.zeebe.broker.client.api.BrokerClient;
-import io.camunda.zeebe.broker.client.api.BrokerClusterState;
-import io.camunda.zeebe.broker.client.api.BrokerTopologyManager;
+import io.camunda.security.auth.CamundaAuthentication;
+import io.camunda.security.auth.CamundaAuthenticationProvider;
+import io.camunda.service.TopologyServices;
+import io.camunda.service.TopologyServices.ClusterStatus;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
-import io.camunda.zeebe.protocol.record.PartitionHealthStatus;
-import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
-import org.mockito.Mock;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -31,31 +28,22 @@ class StatusControllerTest extends RestControllerTest {
 
   static final String STATUS_URL = "/v2/status";
 
-  @MockitoBean BrokerClient brokerClient;
-  @MockitoBean BrokerTopologyManager topologyManager;
-  @Mock BrokerClusterState clusterState;
+  @MockitoBean TopologyServices topologyServices;
+  @MockitoBean CamundaAuthenticationProvider authenticationProvider;
 
   @BeforeEach
   void setUp() {
-    when(brokerClient.getTopologyManager()).thenReturn(topologyManager);
-    when(topologyManager.getTopology()).thenReturn(clusterState);
+    when(authenticationProvider.getCamundaAuthentication())
+        .thenReturn(AUTHENTICATION_WITH_DEFAULT_TENANT);
+    when(topologyServices.withAuthentication(any(CamundaAuthentication.class)))
+        .thenReturn(topologyServices);
   }
 
   @Test
-  void shouldReturnNoContentWhenAtLeastOnePartitionHasHealthyLeader() {
+  void shouldReturnNoContentWhenHealthy() {
     // given
-    final int partitionId1 = 1;
-    final int partitionId2 = 2;
-    final int leaderId1 = 1;
-    final int leaderId2 = 2;
-
-    when(clusterState.getPartitions()).thenReturn(List.of(partitionId1, partitionId2));
-    when(clusterState.getLeaderForPartition(partitionId1)).thenReturn(leaderId1);
-    when(clusterState.getLeaderForPartition(partitionId2)).thenReturn(leaderId2);
-    when(clusterState.getPartitionHealth(leaderId1, partitionId1))
-        .thenReturn(PartitionHealthStatus.HEALTHY);
-    when(clusterState.getPartitionHealth(leaderId2, partitionId2))
-        .thenReturn(PartitionHealthStatus.UNHEALTHY);
+    when(topologyServices.getStatus())
+        .thenReturn(CompletableFuture.completedFuture(ClusterStatus.HEALTHY));
 
     // when / then
     webClient
@@ -67,32 +55,13 @@ class StatusControllerTest extends RestControllerTest {
         .isNoContent()
         .expectBody()
         .isEmpty();
-
-    verify(brokerClient).getTopologyManager();
-    verify(topologyManager).getTopology();
-    verify(clusterState).getPartitions();
-    verify(clusterState).getLeaderForPartition(partitionId1);
-    verify(clusterState).getPartitionHealth(leaderId1, partitionId1);
   }
 
-  @ParameterizedTest
-  @EnumSource(
-      value = PartitionHealthStatus.class,
-      names = {"HEALTHY"},
-      mode = EnumSource.Mode.EXCLUDE)
-  void shouldReturnServiceUnavailableWhenNoPartitionHasHealthyLeader(
-      final PartitionHealthStatus unhealthyStatus) {
+  @Test
+  void shouldReturnServiceUnavailableWhenNoPartitionHasHealthyLeader() {
     // given
-    final int partitionId1 = 1;
-    final int partitionId2 = 2;
-    final int leaderId1 = 1;
-    final int leaderId2 = 2;
-
-    when(clusterState.getPartitions()).thenReturn(List.of(partitionId1, partitionId2));
-    when(clusterState.getLeaderForPartition(partitionId1)).thenReturn(leaderId1);
-    when(clusterState.getLeaderForPartition(partitionId2)).thenReturn(leaderId2);
-    when(clusterState.getPartitionHealth(leaderId1, partitionId1)).thenReturn(unhealthyStatus);
-    when(clusterState.getPartitionHealth(leaderId2, partitionId2)).thenReturn(unhealthyStatus);
+    when(topologyServices.getStatus())
+        .thenReturn(CompletableFuture.completedFuture(ClusterStatus.UNHEALTHY));
 
     // when / then
     webClient
@@ -104,78 +73,5 @@ class StatusControllerTest extends RestControllerTest {
         .isEqualTo(SERVICE_UNAVAILABLE)
         .expectBody()
         .isEmpty();
-
-    verify(brokerClient).getTopologyManager();
-    verify(topologyManager).getTopology();
-    verify(clusterState).getPartitions();
-    verify(clusterState).getLeaderForPartition(partitionId1);
-    verify(clusterState).getPartitionHealth(leaderId1, partitionId1);
-    verify(clusterState).getLeaderForPartition(partitionId2);
-    verify(clusterState).getPartitionHealth(leaderId2, partitionId2);
-  }
-
-  @Test
-  void shouldReturnServiceUnavailableWhenNoPartitionsExist() {
-    // given
-    when(clusterState.getPartitions()).thenReturn(List.of());
-
-    // when / then
-    webClient
-        .get()
-        .uri(STATUS_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .exchange()
-        .expectStatus()
-        .isEqualTo(SERVICE_UNAVAILABLE)
-        .expectBody()
-        .isEmpty();
-
-    verify(brokerClient).getTopologyManager();
-    verify(topologyManager).getTopology();
-    verify(clusterState).getPartitions();
-  }
-
-  @Test
-  void shouldHandleMixedPartitionHealthStatuses() {
-    // given
-    final int partitionId1 = 1;
-    final int partitionId2 = 2;
-    final int partitionId3 = 3;
-    final int leaderId1 = 1;
-    final int leaderId2 = 2;
-    final int leaderId3 = 3;
-
-    when(clusterState.getPartitions())
-        .thenReturn(List.of(partitionId1, partitionId2, partitionId3));
-    when(clusterState.getLeaderForPartition(partitionId1)).thenReturn(leaderId1);
-    when(clusterState.getLeaderForPartition(partitionId2)).thenReturn(leaderId2);
-    when(clusterState.getLeaderForPartition(partitionId3)).thenReturn(leaderId3);
-    when(clusterState.getPartitionHealth(leaderId1, partitionId1))
-        .thenReturn(PartitionHealthStatus.DEAD);
-    when(clusterState.getPartitionHealth(leaderId2, partitionId2))
-        .thenReturn(PartitionHealthStatus.UNHEALTHY);
-    when(clusterState.getPartitionHealth(leaderId3, partitionId3))
-        .thenReturn(PartitionHealthStatus.HEALTHY);
-
-    // when / then
-    webClient
-        .get()
-        .uri(STATUS_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .exchange()
-        .expectStatus()
-        .isNoContent()
-        .expectBody()
-        .isEmpty();
-
-    verify(brokerClient).getTopologyManager();
-    verify(topologyManager).getTopology();
-    verify(clusterState).getPartitions();
-    verify(clusterState).getLeaderForPartition(partitionId1);
-    verify(clusterState).getPartitionHealth(leaderId1, partitionId1);
-    verify(clusterState).getLeaderForPartition(partitionId2);
-    verify(clusterState).getPartitionHealth(leaderId2, partitionId2);
-    verify(clusterState).getLeaderForPartition(partitionId3);
-    verify(clusterState).getPartitionHealth(leaderId3, partitionId3);
   }
 }
